feat: Add email verification dialog with blurred loader
This commit introduces a new feature that displays an email verification dialog using `AppUtils.showDialog()`. With two options `Verify` & `Cancel` On pressing `Verify`, a blurred loader is presented to the user with `AppUtils.showBlurredLoaderDialog()`, ensuring a smoother user experience during email verification and right after that it leads to email verifications.

- also added a new dart class as AppUtils to store all the utility functions

Closes #146